+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# [5.13.4] - 2024-06-11
+
+### Added
+
+* The following report endpoints can override the base URI and authentication headers per request if needed:
+  - `ChangeHealth::Request::Claim::Report.report_list`
+  - `ChangeHealth::Request::Claim::Report.get_report`
+  - `ChangeHealth::Request::Claim::Report.delete_report`
+
+  Provide the following parameters to override the defaults set in Configuration:
+  - `base_uri`
+  - `auth_headers` - an empty hash can also be provided (`{}`), which will issue a request to the authentication endpoint instead of using the configured headers.
+
 # [5.13.3] - 2024-05-20
 
 ### Fixed
@@ -692,6 +705,7 @@ Added the ability to hit professional claim submission API. For more details, se
 * Authentication
 * Configuration
 
+[5.13.4]: https://github.com/WeInfuse/change_health/compare/v5.13.3...v5.13.4
 [5.13.3]: https://github.com/WeInfuse/change_health/compare/v5.13.2...v5.13.3
 [5.13.2]: https://github.com/WeInfuse/change_health/compare/v5.13.1...v5.13.2
 [5.13.1]: https://github.com/WeInfuse/change_health/compare/v5.13.0...v5.13.1